requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery[redis]==5.3.4
redis==5.0.1
python-multipart==0.0.6
docker==6.1.3
paramiko==3.3.1
gitpython==3.1.40
websockets==12.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
